refactor(backend): moved some logic from chargefw2
