+++ new/src/backend/app/api/v1/routes/web/charges.py
@@ -24,6 +24,8 @@ from core.exceptions.http import BadRequestError, NotFoundError
 
 
 from db.repositories.calculation_set_repository import CalculationSetFilters
+from services.calculation_storage import CalculationStorageService
+from services.mmcif import MmCIFService
 from services.io import IOService
 from services.chargefw2 import ChargeFW2Service
 
@@ -127,6 +129,7 @@ async def calculate_charges(
         Literal["charges", "none"], Query(description="Output format.")
     ] = "charges",
     chargefw2: ChargeFW2Service = Depends(Provide[Container.chargefw2_service]),
+    storage_service: CalculationStorageService = Depends(Provide[Container.storage_service]),
 ):
     """
     Calculates partial atomic charges for files in the provided directory.
@@ -139,7 +142,7 @@ async def calculate_charges(
             detail="No configurations provided.",
         )
 
-    if chargefw2.get_calculation_set(computation_id) is None:
+    if storage_service.get_calculation_set(computation_id) is None:
         raise NotFoundError(detail=f"Computation '{computation_id}' not found.")
 
     try:
@@ -166,7 +169,7 @@ async def setup(
     request: Request,
     files: list[UploadFile],
     io: IOService = Depends(Provide[Container.io_service]),
-    chargefw2: ChargeFW2Service = Depends(Provide[Container.chargefw2_service]),
+    storage_service: CalculationStorageService = Depends(Provide[Container.storage_service]),
 ) -> Response[Setup]:
     """Stores the provided files on disk and returns the computation id."""
 
@@ -198,7 +201,7 @@ async def setup(
         workdir = io.get_input_path(computation_id)
         io.create_dir(workdir)
         await asyncio.gather(*[io.store_upload_file(file, workdir) for file in files])
-        chargefw2.store_calculation_set(computation_id, user_id, [])
+        storage_service.store_calculation_set(computation_id, user_id, [])
 
         return Response(data={"computationId": computation_id})
     except Exception as e:
@@ -214,13 +217,13 @@ async def get_mmcif(
     computation_id: Annotated[str, Path(description="UUID of the computation.")],
     molecule: Annotated[str | None, Query(description="Molecule name.")] = None,
     io: IOService = Depends(Provide[Container.io_service]),
-    chargefw2: ChargeFW2Service = Depends(Provide[Container.chargefw2_service]),
+    mmcif_service: MmCIFService = Depends(Provide[Container.mmcif_service]),
 ) -> FileResponse:
     """Returns a mmcif file for the provided molecule in the computation."""
 
     try:
         charges_path = io.get_charges_path(computation_id)
-        mmcif_path = chargefw2.get_molecule_mmcif(charges_path, molecule)
+        mmcif_path = mmcif_service.get_molecule_mmcif(charges_path, molecule)
         return FileResponse(path=mmcif_path)
     except FileNotFoundError as e:
         raise NotFoundError(detail=f"MMCIF file for molecule '{molecule}' not found.") from e
@@ -237,12 +240,12 @@ async def get_example_mmcif(
     example_id: Annotated[str, Path(description="ID of the example.", example="phenols")],
     molecule: Annotated[str | None, Query(description="Molecule name.")] = None,
     io: IOService = Depends(Provide[Container.io_service]),
-    chargefw2: ChargeFW2Service = Depends(Provide[Container.chargefw2_service]),
+    mmcif_service: MmCIFService = Depends(Provide[Container.mmcif_service]),
 ) -> FileResponse:
     """Returns a mmcif file for the provided molecule in the example."""
     try:
         examples_path = io.get_example_path(example_id)
-        mmcif_path = chargefw2.get_molecule_mmcif(examples_path, molecule)
+        mmcif_path = mmcif_service.get_molecule_mmcif(examples_path, molecule)
         return FileResponse(path=mmcif_path)
     except FileNotFoundError as e:
         raise NotFoundError(
@@ -305,7 +308,7 @@ async def get_calculations(
     page_size: Annotated[int, Query(description="Number of items per page.")] = 10,
     order_by: Annotated[Literal["created_at"], Query(description="Order by field.")] = "created_at",
     order: Annotated[Literal["asc", "desc"], Query(description="Order direction.")] = "desc",
-    chargefw2: ChargeFW2Service = Depends(Provide[Container.chargefw2_service]),
+    storage_service: CalculationStorageService = Depends(Provide[Container.storage_service]),
 ) -> Response[PagedList[CalculationSetPreviewDto]]:
     """Returns all calculations stored in the database."""
     user = request.state.user
@@ -320,7 +323,7 @@ async def get_calculations(
         filters = CalculationSetFilters(
             order=order, order_by=order_by, page=page, page_size=page_size, user_id=user.id
         )
-        calculations = chargefw2.get_calculations(filters)
+        calculations = storage_service.get_calculations(filters)
         return Response(data=calculations)
     except Exception as e:
         raise BadRequestError(
